+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Updated
+  * Updated wdio.conf to dynamically update `IP`, when VPN is connected.
+
 * Changed
   * Updated browserslist-config-terra to @cerner/browserslist-config-terra